+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Made `GetCustomRewardRedemptionRequest::reward_id` optional
 - `typed-builder` is now optional and not enabled by default, enable feature `typed-builder` to use the `::builder()` methods
 - Client ext methods that take a `IntoIterator<Item = T>` now takes a `IntoIterator<Item = impl Into<T>>`
+- `Get Channel Information` can now take multiple ids
 
 ### Changes